dom/workers/test/test_dynamicImport_early_termination.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<!--
Tests of Worker Dynamic Import (Bug 1540913)
Ensure that the script loader doesn't fail if requests are terminated early.
-->
<head>
  <title>Test for Worker Dynamic Import (Bug 1540913)</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="onLoad()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1540913">Worker Dynamic Import
 Bug 1540913</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();

async function onLoad() {
  await SpecialPowers.pushPrefEnv(
      { set: [["dom.workers.modules.enabled", true ]] });

  const workers = [
    new Worker("dynamicImport_worker.js", {type: "classic"}),
    new Worker("dynamicImport_worker.js", {type: "module"})
  ]

  let successCount = 0;

  // In the implementation of dynamic import, every dynamic import has
  // it's own ScriptLoader. To ensure that this is working correctly,
  // this tests that if we re-order the dynamic import order,
  // worker termination works as expected.
  for (const worker of workers) {
    const events = [];
    worker.onmessage = function(event) {
      switch (event.data) {
        case "first":
          ok(false, "first dynamic import returned");
          SimpleTest.finish();
          break;
        case "second":
          ok(events.length === 0,
             "second dynamic import returned");
          events.push(event.data);
          worker.terminate()
          successCount++;
          // Cheap way to make sure we only finish successfully after
          // both the module and classic test is finished.
          if (successCount == 2) {
            SimpleTest.finish();
          }
          break;
        default:
          ok(false, "Unexpected message:" + event.data);
          SimpleTest.finish();
      }
    };

    worker.onerror = function(event) {
      ok(false, "Worker had an error:" + event.message);
      SimpleTest.finish();
    }

    worker.postMessage("start");
  }
}
</script>
</pre>
</body>
</html>
